requirements.txt
numpy==1.21.4
deepspeech==0.9.3
pyaudio==0.2.11
requests==2.21.0
flask==1.1.2
Flask-Login==0.5.0
flask_cors==3.0.10
flask-socketio==5.1.0
flask-session==0.4.0
bcrypt==3.2.0
pyttsx3==2.90
py-marytts==0.1.4
python-socketio==5.3.0
rsa==4.7.2
youtube-search-python==1.4.7
python-vlc==3.0.12118
beautifulsoup4==4.9.3
git+https://github.com/Cupcakus/pafy
pixel-ring==0.1.0
pyalsaaudio==0.9.0
youtube_dl==2021.6.6
py-picotts==0.1.2
simpleaudio==1.0.4
requests-ntlm==1.1.0
pyopenssl==21.0.0
